Move StatWindow to util.py rename it to utils.py
